Move utilities for working with Strings in terms of code points
These are going to be used by code other than bootstring, so I'm moving them to the `core` package object.

They also should probably not have been made public, so I'm making them `private[idna4s]` which is a binary breaking change.
